# openapi.yaml
openapi: 3.1.0
info:
  title: IFRS9 API
  version: 1.0.0
paths:
...
  /reports/{portfolio_id}/generate:
    post:
      tags:
      - reports
      summary: Generate Report
      description: Generate various types of reports for a portfolio
      operationId: generate_report_reports__portfolio_id__generate_post
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: portfolio_id
        in: path
        required: true
        schema:
          type: integer
          title: Portfolio Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReportRequest'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '404':
          description: Portfolio not found
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /reports/{portfolio_id}/history:
    get:
      tags:
      - reports
      summary: Get Report History
      description: Get report history for a portfolio with optional filters
      operationId: get_report_history_reports__portfolio_id__history_get
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: portfolio_id
        in: path
        required: true
        schema:
          type: integer
          title: Portfolio Id
      - name: report_type
        in: query
        required: false
        schema:
          anyOf:
          - $ref: '#/components/schemas/ReportTypeEnum'
          - type: 'null'
          title: Report Type
      - name: start_date
        in: query
        required: false
        schema:
          anyOf:
          - type: string
            format: date
          - type: 'null'
          title: Start Date
      - name: end_date
        in: query
        required: false
        schema:
          anyOf:
          - type: string
            format: date
          - type: 'null'
          title: End Date
      - name: skip
        in: query
        required: false
        schema:
          type: integer
          default: 0
          title: Skip
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          default: 20
          title: Limit
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReportHistoryList'
        '404':
          description: Portfolio not found
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /reports/{portfolio_id}/report/{report_id}:
    get:
      tags:
      - reports
      summary: Get Report
      description: Get a specific report by ID
      operationId: get_report_reports__portfolio_id__report__report_id__get
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: portfolio_id
        in: path
        required: true
        schema:
          type: integer
          title: Portfolio Id
      - name: report_id
        in: path
        required: true
        schema:
          type: integer
          title: Report Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReportResponse'
        '404':
          description: Portfolio or Report not found
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    delete:
      tags:
      - reports
      summary: Delete Report
      description: Delete a specific report by ID
      operationId: delete_report_reports__portfolio_id__report__report_id__delete
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: portfolio_id
        in: path
        required: true
        schema:
          type: integer
          title: Portfolio Id
      - name: report_id
        in: path
        required: true
        schema:
          type: integer
          title: Report Id
      responses:
        '204':
          description: Successful Response
        '404':
          description: Portfolio or Report not found
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /reports/{portfolio_id}/report/{report_id}/download:
    get:
      tags:
      - reports
      summary: Download Report Excel
      description: Download a specific report as Excel
      operationId: download_report_excel_reports__portfolio_id__report__report_id__download_get
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: portfolio_id
        in: path
        required: true
        schema:
          type: integer
          title: Portfolio Id
      - name: report_id
        in: path
        required: true
        schema:
          type: integer
          title: Report Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '404':
          description: Portfolio or Report not found
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /reports/status/{report_id}:
    get:
      tags:
      - reports
      summary: Get Report Status
      description: Check status of a report generation
      operationId: get_report_status_reports_status__report_id__get
      parameters:
      - name: report_id
        in: path
        required: true
        schema:
          type: integer
          title: Report Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '404':
          description: Report not found
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    ReportHistoryItem:
      properties:
        id:
          type: integer
          title: Id
        report_type:
          type: string
          title: Report Type
        report_date:
          type: string
          format: date
          title: Report Date
        report_name:
          type: string
          title: Report Name
        created_at:
          type: string
          format: date-time
          title: Created At
      type: object
      required:
      - id
      - report_type
      - report_date
      - report_name
      - created_at
      title: ReportHistoryItem
    ReportHistoryList:
      properties:
        items:
          items:
            $ref: '#/components/schemas/ReportHistoryItem'
          type: array
          title: Items
        total:
          type: integer
          title: Total
      type: object
      required:
      - items
      - total
      title: ReportHistoryList
    ReportRequest:
      properties:
        report_date:
          type: string
          format: date
          title: Report Date
        report_type:
          $ref: '#/components/schemas/ReportTypeEnum'
      type: object
      required:
      - report_date
      - report_type
      title: ReportRequest
    ReportResponse:
      properties:
        report_type:
          $ref: '#/components/schemas/ReportTypeEnum'
        report_date:
          type: string
          format: date
          title: Report Date
        report_name:
          type: string
          title: Report Name
        id:
          type: integer
          title: Id
        portfolio_id:
          type: integer
          title: Portfolio Id
        created_at:
          type: string
          format: date-time
          title: Created At
        created_by:
          type: integer
          title: Created By
        report_data:
          type: object
          title: Report Data
      type: object
      required:
      - report_type
      - report_date
      - report_name
      - id
      - portfolio_id
      - created_at
      - created_by
      - report_data
      title: ReportResponse
    ReportTypeEnum:
      type: string
      enum:
      - ecl_detailed_report
      - ecl_report_summarised_by_stages
      - BOG_impairment_detailed_report
      - BOG_impairmnt_summary_by_stages
      - journals_report
      title: ReportTypeEnum
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
  securitySchemes:
    OAuth2PasswordBearer:
      type: oauth2
      flows:
        password:
          scopes: {}
          tokenUrl: token